requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.10
django-extensions==3.1.5
django-sqlformatter==1.0.3
mysqlclient==2.1.0
Pygments==2.11.2
pytz==2022.1
sqlparse==0.4.2
djangorestframework